refactor: folder and file names
