+++ src/shared/utils/index.ts
@@ -1,2 +1,3 @@
 export { getDefaultTheme } from './get-default-theme';
+export { getRandomMessage } from './get-random-message';
 export { localStorageController } from './local-storage-controller';
