+++ src/widgets/footer.tsx
