+++ src/widgets/header/auth-controls.tsx
