+++ src/widgets/header/logo.tsx
