+++ src/widgets/header/navigation.tsx
